Fix the about page
Before, we'd error on the about page, since it expects
"interview_metadata["main_interview_key"]" to be a string, not the entire
dictionary.

However, much of that code is no longer needed in the latest AssemblyLine
(v2.27.0), which included an upstream fix to use the much simpler `metadata`
section instead.

The resulting about screen works! There might be some information to add, but
we can do that later.

Fix #10, Fix #20